% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores_check.r
\name{calculate_scores_check}
\alias{calculate_scores_check}
\title{Calculate scores for OHI repository}
\usage{
calculate_scores_check(repo_registry, push = FALSE)
}
\arguments{
\item{repo_registry}{data frame with information about the repo}

\item{push}{T/F: do you want to add, commit, and push? Defaults to FALSE.}
}
\value{
TBD
}
\description{
Calculate scores for OHI repository
}
\examples{
\dontrun{# TBD}
}
