% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_and_push.r
\name{commit_and_push}
\alias{commit_and_push}
\title{commit_and_push}
\usage{
commit_and_push(key, dir_repo, commit_message, branch = "master")
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_repo}{local directory where you have cloned the repo (probably somewhere temporary)}

\item{commit_message}{short message in quotes to accompany commit}

\item{branch}{GitHub branch to push, defaults to 'master'}
}
\description{
commit_and_push
}
