% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_layer.r
\name{copy_layer}
\alias{copy_layer}
\title{Copy OHI Layer}
\usage{
copy_layer(lyr = lyr, rgns_key, dir_origin, dir_scenario, lyrs_key,
  write_to_csv = TRUE)
}
\arguments{
\item{lyr}{OHI data layer to be copied}

\item{rgns_key}{regions list for key}

\item{dir_origin}{full local path of origin repo (e.g. ohi-global/eez)}

\item{lyrs_key}{layers.csv data object}

\item{write_to_csv}{whether to write to .csv; default is TRUE}
}
\description{
Copy OHI Layer
}
