% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gh_repo.R
\name{create_gh_repo}
\alias{create_gh_repo}
\title{Create GitHub Repo}
\usage{
create_gh_repo(key, gh_token, github_user)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{gh_token}{GitHub Token}

\item{github_user}{GitHub User, set in create_init.r}
}
\description{
Create GitHub Repo
}

