% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo_map.R
\name{create_repo_map}
\alias{create_repo_map}
\title{Title}
\usage{
create_repo_map(repo_registry, push = TRUE)
}
\arguments{
\item{repo_registry}{data frame with information about the repo}

\item{push}{TRUE/FALSE: do you want to add, commit, and push? Defaults to TRUE.}
}
\value{
repo
}
\description{
Title
}
