% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_website_prep.R
\name{deploy_website_prep}
\alias{deploy_website_prep}
\title{Deploy website to gh-pages branch in Github repository}
\usage{
deploy_website_prep(key, dir_repo, gh_org = "OHI-Science",
  dir_scenario = file.path(dir_repo, repo_registry$scenario_name),
  study_area = repo_registry$study_area, push = TRUE)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_repo}{local directory where you have cloned the repo (probably somewhere temporary)}

\item{gh_org}{github organization to place the repo. Default: ohi-science}

\item{push}{TRUE/FALSE: do you want to add, commit, and push? Defaults to TRUE.}
}
\value{
Returns web_url (http://ohi-science.github.io/\link{key}) based on creating or
updating gh-pages branch of Github repository. Please also visit
\link\link{=https://github.com/OHI-Science/ohirepos/blob/master/inst/gh-pages/README.md}{gh-pages/README.md} for more details.
}
\description{
Deploy website to gh-pages branch in Github repository
}
\examples{
\dontrun{
deploy_website_prep('arc', 'The Arctic')
}

}
