% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_global_region.R
\name{extract_global_region}
\alias{extract_global_region}
\title{Extract global region (without subregions)
Extract global regions, to be used in Tier 1 assessments}
\usage{
extract_global_region(repo_registry, dir_save)
}
\arguments{
\item{repo_registry}{data frame with information about the repo}

\item{save_dir}{directory to save the shapefile}
}
\description{
Extract global region (without subregions)
Extract global regions, to be used in Tier 1 assessments
}
