% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mazu_by_rgn.R
\name{extract_mazu_by_rgn}
\alias{extract_mazu_by_rgn}
\title{Extract global data stored on Mazu for a specified region}
\usage{
extract_mazu_by_rgn(mazu_filename, rgn_name, save_local_dir)
}
\arguments{
\item{mazu_filename}{filepath from MAZU's git-annex}

\item{rgn_name}{region name from OHI Global Assessment}

\item{save_local_dir}{local filepath to save extracted data}
}
\description{
This requires access to the NCEAS MAZU server
}
\examples{
ohicore::extract_mazu_by_rgn(
   mazu_filename = "globalprep/fis/v2017/int/stock_catch_by_rgn.csv",
   rgn_name = "Kenya",
   save_local_dir = "~/github/ken/global_explore/data")


}
