% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_remote_branches.R
\name{gh_remote_branches}
\alias{gh_remote_branches}
\title{Get remote branches on Github from local repository.}
\usage{
gh_remote_branches(dir_repo)
}
\arguments{
\item{dir_repo}{local directory where you have cloned the repo}
}
\value{
Returns character vector of branches
}
\description{
Get remote branches on Github from local repository.
}
