% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_conf.R
\name{populate_conf}
\alias{populate_conf}
\title{Populate OHI repo with configuration folder}
\usage{
populate_conf(key, repo_registry)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{repo_registry}{}
}
\value{
key's repo with config folder populated
}
\description{
Populate OHI repo with configuration folder
}
