% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_etc.r
\name{populate_etc}
\alias{populate_etc}
\title{Populate other items in OHI repos}
\usage{
populate_etc(key, dir_scenario)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_scenario}{full path of temporary OHI repo and scenario, e.g. `~/github/clip-n-ship/gye`}
}
\value{
TBD
}
\description{
Populate other items in OHI repos
}
\examples{
\dontrun{# TBD}
}

