% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_branch.r
\name{push_branch}
\alias{push_branch}
\title{push_branch}
\usage{
push_branch(branch, current_msg)
}
\arguments{
\item{branch}{GitHub branch to push}

\item{current_msg}{Commit message}
}
\description{
push_branch
}
