% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpopulate_layers_conf.r
\name{unpopulate_layers_conf}
\alias{unpopulate_layers_conf}
\title{Unpopulate existing layers and conf folders and files}
\usage{
unpopulate_layers_conf(key, dir_repos, dir_repo, git_url, default_scenario)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_repos}{full path of temporary folder for OHI repos, e.g. \code{~/github/clip-n-ship}}

\item{dir_repo}{full path of temporary OHI repo e.g. \code{~/github/clip-n-ship/gye}}

\item{git_url}{url of key's github repository}

\item{default_scenario}{key's scenario folder}
}
\value{
key's repo after unpopulating layers and conf information
}
\description{
Unpopulate existing layers and conf folders and files
}
\examples{

}
