% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores_check.r
\name{calculate_scores_check}
\alias{calculate_scores_check}
\title{Calculate scores for OHI repository}
\usage{
calculate_scores_check(dir_scenario)
}
\arguments{
\item{dir_scenario}{full path of temporary OHI repo and scenario, e.g. \code{~/github/clip-n-ship/gye}}
}
\value{
TBD
}
\description{
Calculate scores for OHI repository
}
\examples{
\dontrun{# TBD}
}
