% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_repo.r
\name{clone_repo}
\alias{clone_repo}
\title{Clone existing repo}
\usage{
clone_repo(dir_repos, dir_repo, git_url)
}
\arguments{
\item{dir_repos}{TBD}

\item{dir_repo}{TBD}

\item{git_url}{TBD}
}
\value{
TBD
}
\description{
Clone existing repo
}
\examples{
\dontrun{# TBD}
}
