% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_and_push.r
\name{commit_and_push}
\alias{commit_and_push}
\title{commit_and_push}
\usage{
commit_and_push(repo_registry, commit_message, branch = "master")
}
\arguments{
\item{repo_registry}{data frame with information about the repo}

\item{commit_message}{short message in quotes to accompany commit}

\item{branch}{GitHub branch to push, defaults to 'master'}
}
\description{
commit_and_push
}
