% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_layer.r
\name{copy_layer}
\alias{copy_layer}
\title{Copy OHI Layer}
\usage{
copy_layer(lyr, sc_rgns, dir_origin, sfx_global, lyrs_sc, write_to_csv = TRUE)
}
\arguments{
\item{lyr}{OHI data layer to be copied}

\item{sc_rgns}{OHI+ regions (sc = 'subcountry', vestigal)}

\item{dir_origin}{full local path of origin repo (e.g. global)}

\item{sfx_global}{suffix to identify global source for OHI+data layers}

\item{lyrs_sc}{layers.csv data object}

\item{write_to_csv}{whether to write to .csv; default is TRUE}
}
\description{
Copy OHI Layer
}
