% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo_map.R
\name{create_repo_map}
\alias{create_repo_map}
\title{Title}
\usage{
create_repo_map(key, dir_spatial, dir_scenario)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_spatial}{directory of shapefiles on git-annex}

\item{dir_scenario}{OHI scenario directory}
}
\description{
Title
}

