% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R
\name{deploy_app}
\alias{deploy_app}
\title{Deploy Shiny application}
\usage{
deploy_app(gh_repo, app_title, scenario_dirs, gh_owner = "OHI-Science",
  gh_branch_data = "draft", gh_branch_app = "app",
  app_url = sprintf("http://ohi-science.nceas.ucsb.edu/\%s", gh_repo),
  projection = "Mercator", map_shrink_pct = 10, debug = F)
}
\arguments{
\item{gh_repo}{Github repository.}

\item{app_title}{The title for the app, which is typically the study area or place name.}

\item{scenario_dirs}{A character vector of subfolders from the data branch that will be scenarios available for display. The first one will become the default viewed.}

\item{gh_owner}{Github owner. Defaults to "OHI-Science".}

\item{gh_branch_data}{Github branch containing data. Defaults to "draft" and must already exist in the repo.}

\item{gh_branch_app}{Github branch to contain the app. Defaults to "app" and does not have to already exist in the repo.}

\item{app_url}{URL of the application}

\item{projection}{defaults to Mercator, or could be specified as Mollweide, which may be more appropriate for global results.}

\item{map_shrink_pct}{percentage of shrinkage to apply to study area for default map view.}

\item{debug}{produces a Message box with various debug outputs to evaluate reactivity of the app.}
}
\value{
Returns URL of Shiny app if successfully deployed, otherwise errors out.
}
\description{
Deploy Shiny application
}
\examples{
\dontrun{
deploy_app("bhi")
}
}

