% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R, R/deploy_app_generic.R
\name{deploy_app}
\alias{deploy_app}
\alias{deploy_app}
\title{Deploy Shiny application}
\usage{
deploy_app(gh_organization = "OHI-Science", gh_repo, gh_shiny_dir = NULL,
  gh_branch_app = "master",
  app_base_url = "http://ohi-science.nceas.ucsb.edu",
  app_name_remote = NULL, app_server, install_pkgs = FALSE,
  dir_server = "/srv/shiny-server", dir_local = tempdir())

deploy_app(gh_organization = "OHI-Science", gh_repo, gh_shiny_dir = NULL,
  gh_branch_app = "master",
  app_base_url = "http://ohi-science.nceas.ucsb.edu",
  app_name_remote = NULL, app_server, install_pkgs = FALSE,
  dir_server = "/srv/shiny-server", dir_local = tempdir())
}
\arguments{
\item{gh_organization}{Github organization. Defaults to "OHI-Science".}

\item{gh_repo}{Github repository}

\item{gh_branch_app}{Github branch to contain the app. Defaults to "app" and does not have to already exist in the repo.}

\item{app_server}{the secure copy (scp) server location given by username@server:/dir/}

\item{dir_server}{directory on the app_server}

\item{dir_local}{top-level directory to use for populating git repo folder
and branch subfolders within, defaults to tmpdir()}

\item{scenario_dirs}{character vector of subfolders from the data branch that will be scenarios available for display. The first one will become the default viewed.}

\item{study_area}{place name of the entire study area}

\item{gh_owner}{Github owner. Defaults to "OHI-Science".}

\item{gh_branch_data}{Github branch containing data. Defaults to "draft" and must already exist in the repo.}

\item{app_url}{URL of the application}

\item{projection}{defaults to Mercator, or could be specified as Mollweide, which may be more appropriate for global results}

\item{map_shrink_pct}{percentage of shrinkage to apply to study area for default map view}

\item{run_app}{run the Shiny app locally}

\item{open_url}{open the web browser to the app_url}

\item{dir_out}{top-level directory to use for populating git repo folder and branch subfolders within, defaults to tmpdir()}

\item{del_out}{whether to delete output directory when done, defaults to TRUE}

\item{gh_data_commit}{commit of gh_branch_data for freezing the app, ie do not automatically update}

\item{gh_repo}{Github repository}

\item{gh_branch_app}{Github branch that contains the app. Defaults to "master" and
does not have to already exist in the repo.}

\item{app_url}{URL of the application}

\item{app_server}{the secure copy (scp) server location given by username@server:/dir/}

\item{run_app}{run the Shiny app locally}

\item{open_url}{open the web browser to the app_url}

\item{del_out}{whether to delete output directory when done, defaults to TRUE}

\item{dir_server}{directory on the app_server}

\item{gh_data_commit}{commit of gh_branch_data for freezing the app,
ie do not automatically update}
}
\value{
Returns URL of Shiny app if successfully deployed, otherwise errors out. Requires git credentials to push to Github repository,
and SSH keys for secure copying to server.
Suggestions to update Shiny app:
\enumerate{
\item Ensure permissions are set for you and shiny on fitz:
\verb{
 ssh jstewart@fitz
 sudo chown -R jstewart /srv/shiny-server/ohi-global
 sudo chmod -R 775 .
 sudo chgrp -R shiny .
 }
}

Returns URL of Shiny app if successfully deployed, otherwise errors
out. Requires git credentials to push to Github repository,
and SSH keys for secure copying to server.
Suggestions to update Shiny app:
\enumerate{
\item Ensure permissions are set for you and shiny on fitz:
\verb{ssh jstewart@fitz
             sudo chown -R jstewart /srv/shiny-server/ohi-global
             sudo chmod -R 775 .
             sudo chgrp -R shiny .}
\item Copy your repo to dir_local locally. For example, if input
argument \code{dir_local = '~/Desktop/ohirepos_tmp'},
copy \code{'~/github/ohi-global'}
to \code{dir_local = '~/Desktop/ohirepos_tmp/ohi-global/draft'}.
}
Please also visit
\link\link{=https://github.com/OHI-Science/ohirepos/blob/master/inst/app/README.md}{app/README.md}
for more details.
}
\description{
Deploy Shiny application

Deploy Shiny application
}
\examples{
\dontrun{
deploy_app('ohi-global', 'Global', c('eez2015','eez2012','eez2013','eez2014','eez2016'), projection='Mollweide')
deploy_app(       'bhi', 'Baltic', 'baltic2015')
}
\dontrun{
deploy_app('ohi-global', 'Global', c('eez2015','eez2012','eez2013','eez2014','eez2016'),
  projection = 'Mollweide')
deploy_app(       'bhi', 'Baltic', 'baltic2015')
}
}
