% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R
\name{deploy_app}
\alias{deploy_app}
\title{Deploy Shiny application}
\usage{
deploy_app(gh_repo, gh_owner = "OHI-Science")
}
\arguments{
\item{gh_repo}{Github repository.}

\item{gh_owner}{Github owner. Defaults to "OHI-Science"}
}
\value{
Returns URL of app if successfully deployed, otherwise errors out.
}
\description{
Deploy Shiny application
}
\examples{
\dontrun{
deploy_app("bhi")
}
}

