% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_website.R
\name{deploy_website}
\alias{deploy_website}
\title{Deploy website to gh-pages branch in Github repository}
\usage{
deploy_website(gh_repo, study_area, scenario_dir, gh_owner = "OHI-Science",
  gh_branch_data = "draft",
  app_url = sprintf("http://ohi-science.nceas.ucsb.edu/\%s", gh_repo),
  open_url = FALSE, dir_out = tempdir(), del_out = TRUE)
}
\arguments{
\item{gh_repo}{Github repository}

\item{study_area}{place name of the entire study area}

\item{scenario_dir}{subfolder from data branch of repo that will be the default scenario to display on website}

\item{gh_owner}{Github owner. Defaults to "OHI-Science".}

\item{gh_branch_data}{Github branch containing data. Defaults to "draft" and must already exist in the repo.}

\item{app_url}{URL of Shiny application that gets embedded as iframe}

\item{open_url}{open the web browser to the web_url}

\item{dir_out}{top-level directory to use for populating git repo folder and branch subfolders within, defaults to tmpdir()}

\item{del_out}{whether to delete output directory when done, defaults to TRUE}
}
\value{
Returns web_url (http://[gh_owner].github.io/[gh_repo]) based on creating or
updating gh-pages branch of Github repository.
}
\description{
Deploy website to gh-pages branch in Github repository
}
\examples{
\dontrun{
deploy_website('ohi-global', 'Global', 'eez2015')
deploy_website('bhi',        'Baltic', 'baltic2015')
}

}

