% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_website.R
\name{deploy_website}
\alias{deploy_website}
\title{Deploy website to gh-pages branch in Github repository}
\usage{
deploy_website(key, dir_repo, gh_org = "OHI-Science",
  dir_scenario = file.path(dir_repo, repo_registry$scenario_name),
  study_area = repo_registry$study_area, push = TRUE)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_repo}{local directory where you have cloned the repo (probably somewhere temporary)}

\item{gh_org}{github organization to place the repo. Default: ohi-science}

\item{push}{TRUE/FALSE: do you want to add, commit, and push? Defaults to TRUE.}
}
\value{
Returns web_url (http://ohi-science.github.io/\link{key}) based on creating or
updating gh-pages branch of Github repository.
}
\description{
Deploy website to gh-pages branch in Github repository
}
\examples{

}
