% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_website.R
\name{deploy_website}
\alias{deploy_website}
\title{Deploy website to gh-pages branch in Github repository}
\usage{
deploy_website(gh_repo, web_title, scenario_dir, gh_owner = "OHI-Science",
  gh_branch_data = "draft",
  app_url = sprintf("http://ohi-science.nceas.ucsb.edu/\%s", gh_repo),
  open_url = F, debug = F)
}
\arguments{
\item{gh_repo}{Github repository}

\item{web_title}{title for the website, which is typically the study area or place name}

\item{scenario_dir}{subfolder from data branch of repo that will be the default scenario to display on website}

\item{gh_owner}{Github owner. Defaults to "OHI-Science".}

\item{gh_branch_data}{Github branch containing data. Defaults to "draft" and must already exist in the repo.}

\item{app_url}{URL of Shiny application that gets embedded as iframe}

\item{open_url}{open the web browser to the web_url}

\item{debug}{debugging flag}
}
\value{
Returns web_url (http://[gh_owner].github.io/[gh_repo]) based on creating or
updating gh-pages branch of Github repository.
}
\description{
Deploy website to gh-pages branch in Github repository
}
\examples{
\dontrun{
deploy_app('ohi-global', 'Global', 'eez2015')
deploy_app(       'bhi', 'Baltic', 'baltic2015')
}

}

