% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_utils.R
\name{gh_remote_branches}
\alias{gh_remote_branches}
\title{Get remote branches on Github from local repository.}
\usage{
gh_remote_branches(dir_repo)
}
\arguments{
\item{dir_repo}{local directory containing Githubrepository}
}
\value{
Returns character vector of repositories.
}
\description{
Get remote branches on Github from local repository.
}

