% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_conf.R
\name{populate_conf}
\alias{populate_conf}
\title{Populate OHI repo with configuration folder}
\usage{
populate_conf(key, dir_scenario, dir_global)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_scenario}{full path of temporary OHI repo and scenario, e.g. \code{~/github/clip-n-ship/gye/region2015}}

\item{dir_global}{full path of local \code{ohi-global} scenario, e.g. `~/github/ohi-global/eez2016'}
}
\value{
key's repo with config folder populated
}
\description{
Populate OHI repo with configuration folder
}
