% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_etc.r
\name{populate_etc}
\alias{populate_etc}
\title{Populate other items in OHI repos}
\usage{
populate_etc(repo_registry)
}
\arguments{
\item{repo_registry}{data frame with information about the repo}
}
\value{
TBD
}
\description{
Populate other items in OHI repos
}
\examples{
\dontrun{# TBD}
}
