% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_init.r
\name{populate_init}
\alias{populate_init}
\title{Populate initial files including README and .Rproj}
\usage{
populate_init(key, dir_repo, gh_org = "OHI-Science", push = TRUE)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_repo}{local directory where you have cloned the repo (probably somewhere temporary)}

\item{gh_org}{github organization to place the repo. Default: ohi-science}

\item{push}{TRUE/FALSE: do you want to add, commit, and push? Defaults to TRUE.}
}
\description{
First delete any local copy in your directory and clone the repo from github.com/OHI-Science.
}
