% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_layers.r
\name{populate_layers}
\alias{populate_layers}
\title{Populate new OHI repos with layers from global assessments}
\usage{
populate_layers(key, dir_repo, dir_origin,
  lyrs_gl = readr::read_csv(file.path(dir_origin, "eez/layers.csv")),
  dir_scenario, multi_nation = FALSE)
}
\arguments{
\item{key}{OHI assessment identifier, e.g. 'gye' for 'Gulf of Guayaquil'}

\item{dir_repo}{full path of temporary OHI repo e.g. \code{~/github/clip-n-ship/gye}}

\item{dir_origin}{full local path of origin repo (e.g. global)}

\item{lyrs_gl}{list of global layers to copy}

\item{dir_scenario}{full path of temporary OHI repo and scenario, e.g. \code{~/github/clip-n-ship/gye}}

\item{multi_nation}{T/F whether to pull information from multiple nations}
}
\value{
populates OHI repo with layers from global assessments
}
\description{
Populate new OHI repos with layers from global assessments
}
\examples{


}
