% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_prep.R
\name{populate_prep}
\alias{populate_prep}
\title{Populate the repo with prep folder and subfolders and README files}
\usage{
populate_prep(repo_registry, push = TRUE)
}
\arguments{
\item{repo_registry}{data frame with information about the repo}

\item{push}{TRUE/FALSE: do you want to add, commit, and push? Defaults to TRUE.}
}
\description{
Populate the repo with prep folder and subfolders and README files
}
